Add previews to SyncAnotherDevicePromptView; share the rebrand-preview helper
- Add rebranded/legacy #Preview for SyncAnotherDevicePromptView.
- Extract RebrandedPreview + RebrandPreviewOverride into a shared DEBUG-only
  Internal/RebrandedPreview.swift and reuse it from SimplifiedSyncSettingsViewV2
  (dropping the duplicated private copy and its now-unused DesignResourcesKit
  import).
- Move the Desktop-Sync-New-Feature-128 illustration (+ legacy twin) into the
  SyncUI-iOS package and load it via Bundle.module so it renders in previews;
  remove the now-unused copies from the app's SyncAssets.xcassets.
