+++ iOS/LocalPackages/SyncUI-iOS/Sources/SyncUI-iOS/Resources/SyncMedia.xcassets/Desktop-Sync-New-Feature-128.imageset/Contents.json
